% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alspac_toy_data}
\alias{alspac_toy_data}
\title{Little toy dataset for testing explore and clean functions}
\format{
\subsection{\code{alspac_toy_data}}{

A data frame with 500 rows and 500 columns:
\describe{
\item{cidb1234}{Fake mother id}
\item{qlet}{Fake sibling id}
\item{a237, a257, a258, a271, a274...}{Fake random variables}
...
}
}
}
\source{
'{created inhouse to serve as example, see data-raw/alspac_toy_data.R}'
}
\usage{
alspac_toy_data
}
\description{
Little toy dataset for testing explore and clean functions
}
\keyword{datasets}
