% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac_psychometric_scores.R
\name{ccei_score}
\alias{ccei_score}
\title{Compute Crown - Crisp scale anxiety score (CCEI)}
\usage{
ccei_score(sev1, sev2, sev3, asnumeric = FALSE, verbose = TRUE, data = NULL)
}
\arguments{
\item{sev1}{: A vector of variable names with severity 1 (see CCEI manual)}

\item{sev2}{: A vector of variable names with severity 2 (see CCEI manual)}

\item{sev3}{: A vector of variable names with severity 3 (see CCEI manual)}

\item{asnumeric}{: (FALSE) return the score as.numeric. Default behavior returns
a factor because that is the typical structure in ALSPAC data.}

\item{verbose}{: (TRUE) print table detailing the transformation.}

\item{data}{: (NULL) allows to specify the name of the dataset if this is
different than "data".}
}
\value{
The scale score.
}
\description{
Compute Crown - Crisp scale anxiety score (CCEI)
}
\examples{
\dontrun{
data$m_anx_5y <- ccei_score(
  sev1 = c('k3000', 'k3014', 'k3016'),
  sev2 = c('k3002', 'k3005', 'k3011', 'k3019'),
  sev3 = c('k3008'))
}

}
