% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac_dichotomize.R
\name{dichotomize}
\alias{dichotomize}
\title{Dichotomize (and rename) groups of variables}
\usage{
dichotomize(vars, yes = c("1"), no = c("0"), verbose = TRUE, data = NULL)
}
\arguments{
\item{vars}{: A data.frame with the original variable names as comun names and
the name you want to assign to the new variables as frst row (see example)}

\item{yes}{: A vector of labels or values to be recoded as 1}

\item{no}{: A vector of labels or values to be recoded as 0}

\item{verbose}{: (TRUE) print table detailing the transformation.}

\item{data}{: (NULL) allows to specify the name of the dataset if this is
different than "data".}
}
\value{
A data.frame with dichotomized variables
}
\description{
Dichotomize (and rename) groups of variables
}
\examples{
\dontrun{
LE <- dichotomize(
 vars = data.frame('b570'='partner_died', 'b571'='child_died'),
 yes = c('affected a lot','fairly affected','mildly affected','N effect at all'),
 no = c('No')
 )
}

}
