% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac_dichotomize.R
\name{dichotomize_cont}
\alias{dichotomize_cont}
\title{Dichotomize (and rename) groups of continuous variables}
\usage{
dichotomize_cont(
  vars,
  rule = "",
  quantile = FALSE,
  verbose = TRUE,
  data = NULL
)
}
\arguments{
\item{vars}{: A data.frame with the original variable names as comun names and
the name you want to assign to the new variables as frst row (see example)}

\item{rule}{: A string specifying the rule / cutoff for assigning a value of 1.}

\item{quantile}{: (FALSE) whether the cutoff should be specified in quantiles or
original scale (default uses the original variable scale)}

\item{verbose}{: (TRUE) print table detailing the transformation.}

\item{data}{: (NULL) allows to specify the name of the dataset if this is
different than "data".}
}
\value{
A data.frame with dichotomized variables
}
\description{
Dichotomize (and rename) groups of continuous variables
}
\examples{
\dontrun{
np <- dichotomize_cont(
  vars=data.frame('g496'='neighbourhood_problems_21m', 'h366'='neighbourhood_problems_3y'),
  quantile=TRUE)
}
}
