% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac_psychometric_scores.R
\name{epds_score}
\alias{epds_score}
\title{Compute EPDS total depression score}
\usage{
epds_score(set, revset, asnumeric = FALSE, verbose = TRUE, data = NULL)
}
\arguments{
\item{set}{: A vector of variable names coded as (1 = 0) (2 = 1) (3 = 2) (4 = 3)
(see EPDS manual)}

\item{revset}{: A vector of variable names coded as (1 = 3) (2 = 2) (3 = 1) (4 = 0)
(see EPDS manual)}

\item{asnumeric}{: (FALSE) return the score as.numeric. Default behavior returns
a factor because that is the typical structure in ALSPAC data.}

\item{verbose}{: (TRUE) print table detailing the transformation.}

\item{data}{: (NULL) allows to specify the name of the dataset if this is
different than "data".}
}
\value{
The scale score.
}
\description{
Compute EPDS total depression score
}
\examples{
\dontrun{
data$p_dep_5y <- epds_score(
  set = c('ph3030','ph3031','ph3033'),
  revset = c('ph3032','ph3034','ph3035','ph3036','ph3037','ph3038','ph3039'))
}

}
