% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac_explore.R
\name{find_alspac}
\alias{find_alspac}
\alias{f}
\title{Find out which variables are in the dataset.}
\usage{
find_alspac(pref, data = NULL)

f(pref, data = NULL)
}
\arguments{
\item{pref}{: partial string to search for (understands regular expressions).}

\item{data}{: (NULL) allows to specify the name of the dataset if this is
different than "data".}
}
\value{
Column names in \code{data} that contain \code{pref}.
}
\description{
ALSPAC data likely comes in large dataframes containing many columns.
This function helps with quickly finding out which variables are present.
}
\examples{
find_alspac('f01', toy_data) # Specify the data.frame name yourself
f('f01', toy_data)
data <- toy_data
f('f01') # Assumes data is stored in a data.frame called "data"
}
