% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac_explore.R
\name{find_lab}
\alias{find_lab}
\alias{fl}
\title{Search the variable labels are in the dataset.}
\usage{
find_lab(s, data = NULL, case.sensitive = FALSE, to.data.frame = FALSE)

fl(s, data = NULL, case.sensitive = FALSE, to.data.frame = FALSE)
}
\arguments{
\item{s}{: partial string to search for (understands regular expressions).}

\item{data}{: (NULL) allows to specify the name of the dataset if this is
different than "data".}

\item{case.sensitive}{: (FALSE) whther search should be case sensitive.}

\item{to.data.frame}{: (FALSE) whether to return a
dataframe containing the metadata of selected variables.}
}
\value{
Prints variable names and labels that match the search.
}
\description{
This function helps searching through the sea of concepts and measurements available
(based on variable labels).
}
\examples{
find_lab('income')
find_lab('PREG', case.sensitive=TRUE)
fl('insulin')
}
