% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac_explore.R
\name{find_var}
\alias{find_var}
\alias{fv}
\title{Find out which variables are in the dataset.}
\usage{
find_var(
  s,
  data = NULL,
  method = "contains",
  print.labels = TRUE,
  to.data.frame = FALSE
)

fv(
  s,
  data = NULL,
  method = "contains",
  print.labels = TRUE,
  to.data.frame = FALSE
)
}
\arguments{
\item{s}{: partial string to search for (understands regular expressions).}

\item{data}{: (NULL) allows to specify the name of the dataset if this is
different than "data".}

\item{method}{: ("contains") how to search: possible values: "starts", "ends","contains" (default).
Alternative way to control search with out using regular expressions.}

\item{print.labels}{: (TRUE) whether to print the metadata (i.e. variable names, labels and categories)
of selected variables in the console.}

\item{to.data.frame}{: (FALSE; only used when \code{print.labels=TRUE}) whether to return a
dataframe containing the metadata of selected variables.}
}
\value{
Prints column names in \code{data} that contain \code{pref} (and their labels and categories if print.labels==TRUE).
}
\description{
ALSPAC data likely comes in large dataframes containing many columns.
This function helps with quickly finding out which variables are present.
}
\examples{
find_var('f01', toy_data) # Specify the data.frame name yourself
fv('f01', toy_data)
data <- toy_data
fv('f01') # Assumes data is stored in a data.frame called "data"
}
