% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac_clean.R
\name{rm_outliers}
\alias{rm_outliers}
\title{Remove outliers from variable}
\usage{
rm_outliers(var, cutoff = 1.5, verbose = TRUE, data = NULL)
}
\arguments{
\item{var}{: the variable name to be cleaned.}

\item{cutoff}{: (1.5) how many times the IQR should the upper and lower cutoff
be set (the larger the number, the more lenient the outlier removal)}

\item{verbose}{: (TRUE) print out the number of values below and above the cutoff.}

\item{data}{: (NULL) allows to specify the name of the dataset if this is
different than "data".}
}
\value{
The cleaned variable (with outlier values set to NA)
}
\description{
This function sets all values that are outside the range specified to NA.
}
\examples{
data$cheese_clean <- rm_outliers('cheese')
better_data$crackers_clean <- rm_outliers('crackers', data=better_data)
data$unicorn <- rm_outliers('horses', cutoff=5, verbose=FALSE)

}
