% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac_explore.R
\name{sel}
\alias{sel}
\title{Quickly select groups of variables}
\usage{
sel(var.names, times = NULL, sep = "_", data = NULL)
}
\arguments{
\item{var.names}{: string or vector of (partial) variable names to match.}

\item{times}{: (NULL) if variables have repeated measures, this argument is
used to select only specific times. This can be a string or a vector. It
assumes the variable names contain time indication.}

\item{sep}{: ('_') the character separating the time indication in the
variable name (only used if \code{times} is specified)}

\item{data}{: (NULL) allows to specify the name of the dataset if this is
different than "data".}
}
\value{
Vector of variable names that match the search.
}
\description{
Quickly select groups of variables
}
\examples{
sel('depre') # returns, for example c('depre_1','depre_2','depre_3','depresion_total')
sel('depre', times=c(1,3)) # returns, for example c('depre_1','depre_3')

}
