% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Little toy dataset for testing explore and clean functions}
\format{
\subsection{\code{toy_data}}{

A data frame with 5 rows and 3 columns:
\describe{
\item{id}{Fake unique}
\item{f0101, f0102, yikes123, kz021...}{Fake random variables}
...
}
}
}
\source{
'{created inhouse to serve as example}
}
\usage{
toy_data
}
\description{
Little toy dataset for testing explore and clean functions
}
\keyword{datasets}
