% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac_clean.R
\name{make_idc}
\alias{make_idc}
\title{Make a unique child identifier}
\usage{
make_idc(mom.id = "cidb2957", sibling.id = "qlet", data = NULL)
}
\arguments{
\item{mom.id}{: The variable name of the (project specific) mother ID.}

\item{sibling.id}{: (qlet) a two-level variable indicating parity (A or B).}

\item{data}{: (NULL) allows to specify the name of the dataset if this is
different than "data".}
}
\value{
The unique child ID in the numeric form. The last digit is a either a
0 (for first born) or 1 (for second born)
}
\description{
The ALSPAC datasets often lack a unique participant identifier. So let's make
one using information about the mother and sibling status.
}
\examples{
data$IDC <- make_idc(mom.id='yikes123') # Assumes data is stored in a data.frame called "data"
data$IDC <- make_idc(mom.id='yikes123', data=alspac.data) # Specify the data.frame name yourself

}
