% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac_clean.R
\name{make_sex_factor}
\alias{make_sex_factor}
\title{Make a factor with levels 'male' and 'female' from character vector.}
\usage{
make_sex_factor(sex.var = "kz021", data = NULL)
}
\arguments{
\item{sex.var}{: ('kz021') the name of the sex variable.}

\item{data}{: (NULL) allows to specify the name of the dataset if this is
different than "data".}
}
\value{
A factor with levels: 'male' and 'female'.
}
\description{
Make a factor with levels 'male' and 'female' from character vector.
}
\examples{
\dontrun{
data$sex <- make_sex_factor() # Assumes data is stored in a data.frame called "data"
other_data$sex <- make_sex_factor(data=other_data) # Specify the data.frame name yourself
}
}
